+++ app/src/main/java/com/rockspin/bargainbits/utils/DealUtils.java
@@ -14,7 +14,7 @@ import android.support.v4.content.FileProvider;
 import android.text.Html;
 
 import com.rockspin.bargainbits.R;
-import com.rockspin.bargainbits.data.models.cheapshark.AbbreviatedDeal;
+import com.rockspin.bargainbits.data.models.AbbreviatedDeal;
 import com.rockspin.bargainbits.data.models.cheapshark.Deal;
 
 import java.io.File;
@@ -39,9 +39,9 @@ public final class DealUtils {
         final Deal deal = new Deal();
         deal.setDealID(abbreviatedDeal.getDealID());
         deal.setStoreID(abbreviatedDeal.getStoreID());
-        deal.setSalePrice(abbreviatedDeal.getPrice());
-        deal.setNormalPrice(abbreviatedDeal.getRetailPrice());
-        deal.setSavings(abbreviatedDeal.getSavingsFraction() * 100.0f);
+        deal.setSalePrice((float) abbreviatedDeal.getPrice());
+        deal.setNormalPrice((float) abbreviatedDeal.getRetailPrice());
+        deal.setSavings((float) (abbreviatedDeal.getSavingsFraction() * 100.0f));
 
         return deal;
     }
@@ -56,9 +56,9 @@ public final class DealUtils {
         final Deal deal = new Deal();
         deal.setDealID(abbreviatedDeal.getDealID());
         deal.setStoreID(abbreviatedDeal.getStoreID());
-        deal.setSalePrice(abbreviatedDeal.getPrice());
-        deal.setNormalPrice(abbreviatedDeal.getRetailPrice());
-        deal.setSavings(abbreviatedDeal.getSavingsFraction() * 100.0f);
+        deal.setSalePrice((float) abbreviatedDeal.getPrice());
+        deal.setNormalPrice((float) abbreviatedDeal.getRetailPrice());
+        deal.setSavings((float) (abbreviatedDeal.getSavingsFraction() * 100.0f));
 
         return deal;
     }
